requirements.txt
# Core pipeline dependencies
pyspark==3.5.1
kafka-python==2.0.2
psycopg2-binary==2.9.9
sqlalchemy==2.0.30

# HuggingFace dataset loader
datasets==2.19.1
huggingface-hub==0.23.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.1

# Data processing
pandas==2.2.2
numpy==1.26.4
pyarrow==16.1.0

# Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-mock==3.14.0

# Logging & observability
structlog==24.2.0

# Type checking
mypy==1.10.0
